resturant/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-allauth==65.9.0
gunicorn==23.0.0
packaging==25.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
